+++ src/sintaxtree/Nodes/Node.java
@@ -2,5 +2,5 @@
 package sintaxtree.Nodes;
 
 public abstract class Node {    
-    public abstract int evaluate();
+    public abstract double evaluate();
 }
