+++ src/sintaxtree/Nodes/OperatiorNode.java
@@ -4,5 +4,5 @@ package sintaxtree.Nodes;
 
 public abstract class OperatiorNode extends Node{
     @Override
-    public abstract int evaluate();
+    public abstract double evaluate();
 }
